Fix device tests, move mtop elasticity_pa_regression to test_elasticity_pa unit test, add missing headers
