+++ miniapps/mtop/CMakeLists.txt
@@ -9,10 +9,6 @@
 # terms of the BSD-3 license. We welcome feedback and contributions, see file
 # CONTRIBUTING.md for details.
 
-add_mfem_miniapp(elasticity_pa_regression
-				 MAIN elasticity_pa_regression.cpp
-				 LIBRARIES mfem)
-
 # Skip dFEM qfunction with floats
 if(MFEM_USE_MPI AND MFEM_USE_DOUBLE)
