+++ miniapps/mtop/mtop_ex_linear_elasticity_solver.cpp
@@ -1,4 +1,13 @@
-// Copyright (c) 2010-2025, Lawrence Livermore National Security, LLC.
+// Copyright (c) 2010-2025, Lawrence Livermore National Security, LLC. Produced
+// at the Lawrence Livermore National Laboratory. All Rights reserved. See files
+// LICENSE and NOTICE for details. LLNL-CODE-806117.
+//
+// This file is part of the MFEM library. For more information and source code
+// availability visit https://mfem.org.
+//
+// MFEM is free software; you can redistribute it and/or modify it under the
+// terms of the BSD-3 license. We welcome feedback and contributions, see file
+// CONTRIBUTING.md for details.
 //
 // Driver coverage:
 //   1. PA elasticity solve and inverse-operator consistency.
